[refactor] Fix warnings

+++ gaemi/ImRenderer.h
@@ -15,9 +15,9 @@ namespace gmath {
 using gmath::Rect;
 
 namespace render::im {
-    GAPI f32 GuiSlider(const Rect& rect, const str& textLeft, const str& textRight, f32 value, f32 minValue, f32 maxValue);
-    GAPI f32 GuiSliderBar(const Rect& rect, const str& textLeft, const str& textRight, f32 value, f32 minValue, f32 maxValue);
-    GAPI f32 GuiProgressBar(const Rect& rect, const str& textLeft, const str& textRight, f32 value, f32 minValue, f32 maxValue);
+    GAPI i32 GuiSlider(const Rect& rect, const str& textLeft, const str& textRight, f32 value, f32 minValue, f32 maxValue);
+    GAPI i32 GuiSliderBar(const Rect& rect, const str& textLeft, const str& textRight, f32 value, f32 minValue, f32 maxValue);
+    GAPI i32 GuiProgressBar(const Rect& rect, const str& textLeft, const str& textRight, f32 value, f32 minValue, f32 maxValue);
 }
 
 #endif //GAEMI_IM_RENDERER_H
